+++ cpp/libjoynrclustercontroller/include/joynr/ClusterControllerSettings.h
@@ -38,6 +38,7 @@ public:
     static const std::string& SETTING_LOCAL_CAPABILITIES_DIRECTORY_PERSISTENCY_ENABLED();
     static const std::string& SETTING_LOCAL_DOMAIN_ACCESS_STORE_PERSISTENCE_FILENAME();
     static const std::string& SETTING_MESSAGE_QUEUE_LIMIT();
+    static const std::string& SETTING_PER_PARTICIPANTID_MESSAGE_QUEUE_LIMIT();
     static const std::string& SETTING_MQTT_CLIENT_ID_PREFIX();
     static const std::string& SETTING_MQTT_TLS_ENABLED();
     static const std::string& SETTING_MQTT_CERTIFICATE_AUTHORITY_PEM_FILENAME();
@@ -76,6 +77,7 @@ public:
     static bool DEFAULT_USE_ONLY_LDAS();
     static bool DEFAULT_ACCESS_CONTROL_AUDIT();
     static std::uint64_t DEFAULT_MESSAGE_QUEUE_LIMIT();
+    static std::uint64_t DEFAULT_PER_PARTICIPANTID_MESSAGE_QUEUE_LIMIT();
 
     explicit ClusterControllerSettings(Settings& settings);
     ClusterControllerSettings(const ClusterControllerSettings&) = default;
@@ -126,6 +128,9 @@ public:
 
     std::uint64_t getMessageQueueLimit() const;
     void setMessageQueueLimit(std::uint64_t limit);
+
+    std::uint64_t getPerParticipantIdMessageQueueLimit() const;
+    void setPerParticipantIdMessageQueueLimit(std::uint64_t limit);
 
     bool enableAccessController() const;
     void setEnableAccessController(bool enable);